requirements.txt
phue
speechrecognition
pyaudio
pyttsx3
python-dotenv
androidtv
google-genai
openai
protobuf
requests
python-miio==0.6.0.dev0
plexapi
